% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{average_max_depth_bart_generalized}
\alias{average_max_depth_bart_generalized}
\title{Return the average max depth of all trees and all ensembles in a container of samples}
\usage{
average_max_depth_bart_generalized(bart_result)
}
\arguments{
\item{bart_result}{External pointer to a bart result object}
}
\value{
Average maximum depth
}
\description{
Return the average max depth of all trees and all ensembles in a container of samples
}
