% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{compute_shapley_all}
\alias{compute_shapley_all}
\title{Optimized Global Shapley Calculator}
\usage{
compute_shapley_all(X, beta_post, beta_int_post, ipairs, indices = NULL)
}
\description{
Optimized Global Shapley Calculator
}
