% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{computeForestKernels}
\alias{computeForestKernels}
\title{Compute a kernel from a tree ensemble, defined by the fraction
of trees of an ensemble in which two observations fall into the
same leaf.}
\usage{
computeForestKernels(bart_model, X_train, X_test = NULL, forest_num = NULL)
}
\arguments{
\item{bart_model}{Object of type \code{bartmodel} corresponding to a BART model with at least one sample}

\item{X_train}{"Training" dataframe. In a traditional Gaussian process kriging context, this
corresponds to the observations for which outcomes are observed.}

\item{X_test}{(Optional) "Test" dataframe. In a traditional Gaussian process kriging context, this
corresponds to the observations for which outcomes are unobserved and must be estimated
based on the kernels k(X_test,X_test), k(X_test,X_train), and k(X_train,X_train). If not provided,
this function will only compute k(X_train, X_train).}

\item{forest_num}{(Option) Index of the forest sample to use for kernel computation. If not provided,
this function will use the last forest.}
}
\value{
List of kernel matrices. If \code{X_test = NULL}, the list contains
one \code{n_train} x \code{n_train} matrix, where \code{n_train = nrow(X_train)}.
This matrix is the kernel defined by \code{W_train \%*\% t(W_train)} where \code{W_train}
is a matrix with \code{n_train} rows and as many columns as there are total leaves in an ensemble.
If \code{X_test} is not \code{NULL}, the list contains two more matrices defined by
\code{W_test \%*\% t(W_train)} and \code{W_test \%*\% t(W_test)}.
}
\description{
Compute a kernel from a tree ensemble, defined by the fraction
of trees of an ensemble in which two observations fall into the
same leaf.
}
