% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{convertBARTStateToJson}
\alias{convertBARTStateToJson}
\title{Convert in-memory BART model objects (forests, random effects, vectors) to in-memory JSON.
This function is primarily a convenience function for serialization / deserialization in a parallel BART sampler.}
\usage{
convertBARTStateToJson(
  param_list,
  mean_forest = NULL,
  variance_forest = NULL,
  rfx_samples = NULL,
  global_variance_samples = NULL,
  local_variance_samples = NULL
)
}
\arguments{
\item{param_list}{List containing high-level model state parameters}

\item{mean_forest}{Container of conditional mean forest samples (optional). Default: \code{NULL}.}

\item{variance_forest}{Container of conditional variance forest samples (optional). Default: \code{NULL}.}

\item{rfx_samples}{Container of random effect samples (optional). Default: \code{NULL}.}

\item{global_variance_samples}{Vector of global error variance samples (optional). Default: \code{NULL}.}

\item{local_variance_samples}{Vector of leaf scale samples (optional). Default: \code{NULL}.}
}
\value{
Object of type \code{CppJson}
}
\description{
Convert in-memory BART model objects (forests, random effects, vectors) to in-memory JSON.
This function is primarily a convenience function for serialization / deserialization in a parallel BART sampler.
}
