% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_column_vector}
\alias{create_column_vector}
\title{Create a StochTree C++ ColumVector object (most commonly this is used for the outcome variable)}
\usage{
create_column_vector(vector)
}
\arguments{
\item{vector}{Vector of values}
}
\value{
External pointer to the ColumnVector object
}
\description{
Create a StochTree C++ ColumVector object (most commonly this is used for the outcome variable)
}
\examples{
n <- 100
x <- runif(n)
y <- x*10 + rnorm(n, 0, 0.5)
outcome_ptr <- create_column_vector(y)
}
