% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_forest_dataset}
\alias{create_forest_dataset}
\title{Create a StochTree dataset used for sampling forests}
\usage{
create_forest_dataset(covariates, basis = NULL, weights = NULL)
}
\arguments{
\item{covariates}{Matrix of covariates used for the tree splits}

\item{basis}{(Optional) matrix of bases used to define a regression model in the tree leaves}

\item{weights}{(Optional) vector of variance weights}
}
\value{
External pointer to the dataset
}
\description{
Create a StochTree dataset used for sampling forests
}
\examples{
n <- 100
p_X <- 10
p_W <- 1
X <- matrix(runif(n*p_X), ncol = p_X)
W <- matrix(runif(n*p_W), ncol = p_W)
data_ptr <- create_forest_dataset(X, W)
}
