% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{createBCFModelFromCombinedJsonString}
\alias{createBCFModelFromCombinedJsonString}
\title{Convert a list of (in-memory) JSON strings that represent BART models to a single combined BART model object
which can be used for prediction, etc...}
\usage{
createBCFModelFromCombinedJsonString(json_string_list)
}
\arguments{
\item{json_string_list}{List of JSON strings which can be parsed to objects of type \code{CppJson} containing Json representation of a BART model}
}
\value{
Object of type \code{bartmodel}
}
\description{
Convert a list of (in-memory) JSON strings that represent BART models to a single combined BART model object
which can be used for prediction, etc...
}
\examples{
n <- 100
p <- 5
X <- matrix(runif(n*p), ncol = p)
f_XW <- (
    ((0 <= X[,1]) & (0.25 > X[,1])) * (-7.5) + 
    ((0.25 <= X[,1]) & (0.5 > X[,1])) * (-2.5) + 
    ((0.5 <= X[,1]) & (0.75 > X[,1])) * (2.5) + 
    ((0.75 <= X[,1]) & (1 > X[,1])) * (7.5)
)
noise_sd <- 1
y <- f_XW + rnorm(n, 0, noise_sd)
test_set_pct <- 0.2
n_test <- round(test_set_pct*n)
n_train <- n - n_test
test_inds <- sort(sample(1:n, n_test, replace = FALSE))
train_inds <- (1:n)[!((1:n) \%in\% test_inds)]
X_test <- X[test_inds,]
X_train <- X[train_inds,]
y_test <- y[test_inds]
y_train <- y[train_inds]
bcf_model <- bcf(X_train = X_train, Z_train = Z_train, y_train = y_train, 
                 pi_train = pi_train, group_ids_train = group_ids_train, 
                 rfx_basis_train = rfx_basis_train, X_test = X_test, 
                 Z_test = Z_test, pi_test = pi_test, group_ids_test = group_ids_test,
                 rfx_basis_test = rfx_basis_test, 
                 num_gfr = 100, num_burnin = 0, num_mcmc = 100)
# bcf_json_string_list <- list(saveBCFModelToJsonString(bcf_model))
# bcf_model_roundtrip <- createBCFModelFromCombinedJsonString(bcf_json_string_list)
}
