% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{createCppJson}
\alias{createCppJson}
\title{Create CppJson Object}
\usage{
createCppJson()
}
\value{
\code{CppJson} object
}
\description{
Create a new (empty) C++ Json object
}
\examples{
example_vec <- runif(10)
example_json <- createCppJson()
example_json$add_vector("myvec", example_vec)
}
