% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{createCppJsonFile}
\alias{createCppJsonFile}
\title{Create a C++ Json object from a Json file}
\usage{
createCppJsonFile(json_filename)
}
\value{
\code{CppJson} object
}
\description{
Create a C++ Json object from a Json file
}
