% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createCppRNG}
\alias{createCppRNG}
\title{Create an R class that wraps a C++ random number generator}
\usage{
createCppRNG(random_seed = -1)
}
\arguments{
\item{random_seed}{(Optional) random seed for sampling}
}
\value{
\code{CppRng} object
}
\description{
Create an R class that wraps a C++ random number generator
}
\examples{
rng <- createCppRNG(1234)
rng <- createCppRNG()
}
