% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{createForestKernel}
\alias{createForestKernel}
\title{Create a \code{ForestKernel} object}
\usage{
createForestKernel()
}
\value{
\code{ForestKernel} object
}
\description{
Create a \code{ForestKernel} object
}
