% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{createGlobalModelConfig}
\alias{createGlobalModelConfig}
\title{Create GlobalModelConfig Object}
\usage{
createGlobalModelConfig(global_error_variance = 1)
}
\arguments{
\item{global_error_variance}{Global error variance parameter (default: \code{1.0})}
}
\value{
GlobalModelConfig object
}
\description{
Create a global model config object

This function is intended for advanced use cases in which users require detailed control of sampling algorithms and data structures.
Minimal input validation and error checks are performed -- users are responsible for providing the correct inputs.
For tutorials on the "proper" usage of the stochtree's advanced workflow, we provide several vignettes at stochtree.ai
}
\examples{
config <- createGlobalModelConfig(global_error_variance = 100)
}
