% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createRNG}
\alias{createRNG}
\title{Create an R class that wraps a C++ random number generator}
\usage{
createRNG(random_seed = -1)
}
\arguments{
\item{random_seed}{(Optional) random seed for sampling}
}
\value{
\code{CppRng} object
}
\description{
Create an R class that wraps a C++ random number generator
}
