% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataset_update_basis}
\alias{dataset_update_basis}
\title{Update basis in a StochTree dataset}
\usage{
dataset_update_basis(data_ptr, basis)
}
\arguments{
\item{basis}{Matrix of bases used to define a regression model in the tree leaves}
}
\description{
Update basis in a StochTree dataset
}
\examples{
n <- 100
p_X <- 10
p_W <- 1
X <- matrix(runif(n*p_X), ncol = p_X)
W <- matrix(runif(n*p_W), ncol = p_W)
data_ptr <- create_forest_dataset(X, W)
W_2 <- 2*W
dataset_update_basis(W_2)
}
