% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_dims_1d}
\alias{expand_dims_1d}
\title{Convert scalar input to vector of dimension \code{output_size},
or check that input array is equivalent to a vector of dimension \code{output_size}.}
\usage{
expand_dims_1d(input, output_size)
}
\arguments{
\item{input}{Input to be converted to a vector (or passed through as-is)}

\item{output_size}{Intended size of the output vector}
}
\value{
A vector of length \code{output_size}
}
\description{
Convert scalar input to vector of dimension \code{output_size},
or check that input array is equivalent to a vector of dimension \code{output_size}.
}
