% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_dims_2d_diag}
\alias{expand_dims_2d_diag}
\title{Convert scalar input to square matrix of dimension \code{output_size} x \code{output_size} with \code{input} along the diagonal,
or check that input array is equivalent to a square matrix of dimension \code{output_size} x \code{output_size}.}
\usage{
expand_dims_2d_diag(input, output_size)
}
\arguments{
\item{input}{Input to be converted to a square matrix (or passed through as-is)}

\item{output_size}{Intended row and column dimension of the square output matrix}
}
\value{
A square matrix of dimension \code{output_size} x \code{output_size}
}
\description{
Convert scalar input to square matrix of dimension \code{output_size} x \code{output_size} with \code{input} along the diagonal,
or check that input array is equivalent to a square matrix of dimension \code{output_size} x \code{output_size}.
}
