% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{forest_container}
\alias{forest_container}
\title{Create a container for forest samples}
\usage{
forest_container(num_trees, output_dimension = 1, is_leaf_constant = T)
}
\arguments{
\item{num_trees}{Number of trees}

\item{output_dimension}{Dimensionality of the outcome model}

\item{is_leaf_constant}{Whether leaf is constant}
}
\value{
External pointer to forest container object
}
\description{
Create a container for forest samples
}
\examples{
forest_samples_ptr <- forest_container(100, 1, T)
}
