% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{forest_tracker}
\alias{forest_tracker}
\title{Create a StochTree forest tracker object}
\usage{
forest_tracker(dataset, feature_types, num_trees, n)
}
\arguments{
\item{dataset}{Pointer to a stochtree dataset}

\item{feature_types}{Vector of integers mapping to feature types (0 = numeric, 1 = ordered categorical, 2 = unordered categorical)}

\item{num_trees}{Number of trees in the model}

\item{n}{Size of the training dataset}
}
\value{
External pointer to forest tracker object
}
\description{
Create a StochTree forest tracker object
}
\examples{
n <- 100
p_X <- 10
p_W <- 1
X <- matrix(runif(n*p_X), ncol = p_X)
W <- matrix(runif(n*p_W), ncol = p_W)
data_ptr <- dataset(X, W)
feature_types <- rep(0, p_X)
gfr_sampler_ptr <- gfr_forest_sampler(data_ptr, feature_types, 100, n)
}
