% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{ForestKernel}
\alias{ForestKernel}
\title{Class that provides functionality for statistical kernel definition and
computation based on shared leaf membership of observations in a tree ensemble.}
\description{
Computes leaf membership internally as a sparse matrix and also calculates a
(dense) kernel based on the sparse matrix all in C++.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forest_kernel_ptr}}{External pointer to a C++ StochTree::ForestKernel class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForestKernel-new}{\code{ForestKernel$new()}}
\item \href{#method-ForestKernel-compute_leaf_indices}{\code{ForestKernel$compute_leaf_indices()}}
\item \href{#method-ForestKernel-compute_kernel}{\code{ForestKernel$compute_kernel()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestKernel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForestKernel-new}{}}}
\subsection{Method \code{new()}}{
Create a new ForestKernel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestKernel$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ForestKernel} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestKernel-compute_leaf_indices"></a>}}
\if{latex}{\out{\hypertarget{method-ForestKernel-compute_leaf_indices}{}}}
\subsection{Method \code{compute_leaf_indices()}}{
Compute the leaf indices of each tree in the ensemble for every observation in a dataset.
Stores the result internally, which can be extracted from the class via a call to \code{get_leaf_indices}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestKernel$compute_leaf_indices(
  covariates_train,
  covariates_test = NULL,
  forest_container,
  forest_num
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates_train}}{Matrix of training set covariates at which to compute leaf indices}

\item{\code{covariates_test}}{(Optional) Matrix of test set covariates at which to compute leaf indices}

\item{\code{forest_container}}{Object of type \code{ForestSamples}}

\item{\code{forest_num}}{Index of the forest in forest_container to be assessed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of vectors. If \code{covariates_test = NULL} the list has one element (train set leaf indices), and
otherwise the list has two elements (train and test set leaf indices).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestKernel-compute_kernel"></a>}}
\if{latex}{\out{\hypertarget{method-ForestKernel-compute_kernel}{}}}
\subsection{Method \code{compute_kernel()}}{
Compute the kernel implied by a tree ensemble. This function calls \code{compute_leaf_indices},
so it is not necessary to call both. \code{compute_leaf_indices} is exposed at the class level
to allow for extracting the vector of leaf indices for an ensemble directly in R.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestKernel$compute_kernel(
  covariates_train,
  covariates_test = NULL,
  forest_container,
  forest_num
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariates_train}}{Matrix of training set covariates at which to assess ensemble kernel}

\item{\code{covariates_test}}{(Optional) Matrix of test set covariates at which to assess ensemble kernel}

\item{\code{forest_container}}{Object of type \code{ForestSamples}}

\item{\code{forest_num}}{Index of the forest in forest_container to be assessed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of matrices. If \code{covariates_test = NULL}, the list contains
one \code{n_train} x \code{n_train} matrix, where \code{n_train = nrow(covariates_train)}.
This matrix is the kernel defined by \code{W_train \%*\% t(W_train)} where \code{W_train}
is a matrix with \code{n_train} rows and as many columns as there are total leaves in an ensemble.
If \code{covariates_test} is not \code{NULL}, the list contains two more matrices defined by
\code{W_test \%*\% t(W_train)} and \code{W_test \%*\% t(W_test)}.
}
}
}
