% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_sampler.R
\name{get_initial_values}
\alias{get_initial_values}
\title{Get Initial Values for the Horseshoe Sampler}
\usage{
get_initial_values(X, y, z, p_int_mu, p_int_tau)
}
\arguments{
\item{X}{A numeric matrix of covariates.}

\item{y}{A numeric vector of outcomes.}

\item{z}{A numeric vector of treatment assignments.}

\item{p_int_mu}{Number of interaction terms in the mu model.}

\item{p_int_tau}{Number of interaction terms in the tau model.}
}
\value{
A list of initial values for the sampler.
}
\description{
A helper function to generate reasonable, data-driven starting
values for the MCMC sampler by fitting a single joint linear model.
}
\keyword{internal}
