% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{get_shapley_posterior_dissected}
\alias{get_shapley_posterior_dissected}
\title{Extract Full 5000-Draw Shapley Posterior Dissected into Main and Interaction Effects}
\usage{
get_shapley_posterior_dissected(
  feature_name,
  indices,
  X,
  beta_post,
  beta_int_post,
  ipairs
)
}
\description{
Extract Full 5000-Draw Shapley Posterior Dissected into Main and Interaction Effects
}
