% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{get_shapley_posterior_single}
\alias{get_shapley_posterior_single}
\title{Extract Full 5000-Draw Shapley Posterior for a Single Feature (For Tolerance Intervals)}
\usage{
get_shapley_posterior_single(
  feature_name,
  indices,
  X,
  beta_post,
  beta_int_post,
  ipairs
)
}
\description{
Extract Full 5000-Draw Shapley Posterior for a Single Feature (For Tolerance Intervals)
}
