% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{getRandomEffectSamples.bcfmodel}
\alias{getRandomEffectSamples.bcfmodel}
\title{Extract Random Effects Samples from BCF Model (legacy alias)}
\usage{
\method{getRandomEffectSamples}{bcfmodel}(object, ...)
}
\arguments{
\item{object}{Object of type \code{bcfmodel} containing draws of a BCF model and associated sampling outputs.}

\item{...}{Other parameters to be used in random effects extraction}
}
\value{
List of random effect samples (see \code{\link[=extractRandomEffectSamples]{extractRandomEffectSamples()}}).
}
\description{
Legacy alias for \code{\link[=extractRandomEffectSamples]{extractRandomEffectSamples()}}; delegates to it.
}
\examples{
n <- 100
p <- 5
X <- matrix(runif(n*p), ncol = p)
pi_x <- 0.25 + 0.5*X[,1]
Z <- rbinom(n, 1, pi_x)
rfx_group_ids <- sample(1:2, size = n, replace = TRUE)
rfx_basis <- rep(1.0, n)
mu_x <- X[,1]*2
tau_x <- X[,2]*(-1)
y <- mu_x + tau_x*Z + (-2*(rfx_group_ids==1)+2*(rfx_group_ids==2)) + rnorm(n)
bcf_model <- bcf(X_train=X, Z_train=Z, y_train=y, propensity_train=pi_x,
                 rfx_group_ids_train=rfx_group_ids, rfx_basis_train=rfx_basis,
                 num_gfr=0, num_mcmc=10)
rfx_samples <- getRandomEffectSamples(bcf_model)
}
