% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getRandomEffectSamples}
\alias{getRandomEffectSamples}
\title{Generic function for extracting random effect samples from a model object (BCF, BART, etc...)}
\usage{
getRandomEffectSamples(object, ...)
}
\value{
List of random effect samples
}
\description{
Generic function for extracting random effect samples from a model object (BCF, BART, etc...)
}
