% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{interaction_pairs}
\alias{interaction_pairs}
\title{Generate Interaction Pairs Based on a Boolean Vector
(Fallback function if the model does not provide them natively)}
\usage{
interaction_pairs(num_covariates, boolean_vector)
}
\description{
Generate Interaction Pairs Based on a Boolean Vector
(Fallback function if the model does not provide them natively)
}
