% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{loadRandomEffectSamplesCombinedJsonString}
\alias{loadRandomEffectSamplesCombinedJsonString}
\title{Combine multiple JSON strings representing model objects containing random effects (with the same hierarchy / schema) into a single container}
\usage{
loadRandomEffectSamplesCombinedJsonString(json_string_list, json_rfx_num)
}
\arguments{
\item{json_string_list}{List of objects of class \code{CppJson}}

\item{json_rfx_num}{Integer index indicating the position of the random effects term to be unpacked}
}
\value{
\code{RandomEffectSamples} object
}
\description{
Combine multiple JSON strings representing model objects containing random effects (with the same hierarchy / schema) into a single container
}
