% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oneHotEncode}
\alias{oneHotEncode}
\title{Convert a vector of unordered categorical data (either numeric or character
labels) to a "one-hot" encoded matrix in which a 1 in a column indicates
the presence of the relevant category.}
\usage{
oneHotEncode(x_input, unique_levels)
}
\arguments{
\item{x_input}{Vector of unordered categorical data (typically either strings
integers, but this function also accepts floating point data).}

\item{unique_levels}{Unique values of the categorical variable used to create
the initial one-hot matrix (typically a training set)}
}
\value{
Binary one-hot matrix
}
\description{
This procedure assumes that a reference set of observations for this variable
(typically a training set that was used to sample a forest) has already been
one-hot encoded and that the unique levels of the training set variable are
available (and passed as \code{unique_levels}). Test set observations that contain
categories not in \code{unique_levels} will all be mapped to the last column of
this matrix
}
\examples{
x <- sample(1:8, 100, TRUE)
x_test <- sample(1:9, 10, TRUE)
x_onehot <- oneHotEncode(x_test, levels(factor(x)))
}
