% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orderedCatPreprocess}
\alias{orderedCatPreprocess}
\title{Run some simple preprocessing of ordered categorical variables, converting
ordered levels to integers if necessary, and storing the unique levels of a
variable.}
\usage{
orderedCatPreprocess(x_input, unique_levels, var_name = NULL)
}
\arguments{
\item{x_input}{Vector of ordered categorical data. If the data is not already
stored as an ordered factor, it will be converted to one using the default
sort order.}

\item{unique_levels}{Vector of unique levels for a categorical feature.}

\item{var_name}{(Optional) Name of variable.}
}
\value{
List containing a preprocessed vector of integer-converted ordered
categorical observations and the unique level of the original ordered
categorical feature.
}
\description{
Run some simple preprocessing of ordered categorical variables, converting
ordered levels to integers if necessary, and storing the unique levels of a
variable.
}
\examples{
x_levels <- c("1. Strongly disagree", "2. Disagree", "3. Neither agree nor disagree", 
              "4. Agree", "5. Strongly agree")
x <- c("1. Strongly disagree", "3. Neither agree nor disagree", "2. Disagree", 
       "4. Agree", "3. Neither agree nor disagree", "5. Strongly agree", "4. Agree")
x_processed <- orderedCatPreprocess(x, x_levels)
}
