% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OutcomeModel}
\alias{OutcomeModel}
\title{Create a new outcome model object}
\usage{
OutcomeModel(outcome = "continuous", link = NULL)
}
\arguments{
\item{outcome}{Character string specifying the outcome type.}

\item{link}{Character string specifying the link function.}
}
\value{
An object of class \code{outcome_model}.
}
\description{
Create a new outcome model object
}
\examples{
my_model <- OutcomeModel(outcome = "continuous", link = "identity")
}
