% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{plot.bartmodel}
\alias{plot.bartmodel}
\title{Plot BART Model Fit.}
\usage{
\method{plot}{bartmodel}(x, ...)
}
\arguments{
\item{x}{The BART model object}

\item{...}{Additional arguments}
}
\value{
BART model object unchanged after summarizing
}
\description{
Plot the BART model fit and any relevant sampled quantities. This will default to a traceplot of the global error scale and the in-sample mean forest predictions for the first train set observation. Since \code{stochtree::bart()} is flexible and it's possible to sample a model with a fixed global error scale and no mean forest, this procedure is adaptive and will attempt to plot a trace of whichever model terms are included if these two default terms are omitted.
}
