% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{plot.bcfmodel}
\alias{plot.bcfmodel}
\title{Plot BCF Model}
\usage{
\method{plot}{bcfmodel}(x, ...)
}
\arguments{
\item{x}{The BCF model object}

\item{...}{Additional arguments}
}
\value{
BCF model object unchanged after summarizing
}
\description{
Plot the BCF model fit and any relevant sampled quantities. This will default to a traceplot of the global error scale and the in-sample mean forest predictions for the first train set observation. Since \code{stochtree::bcf()} is flexible and it's possible to sample a model with a fixed global error scale and no mean forest, this procedure will throw an error if these two default terms are omitted.
}
