% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{plot_tolerance_intervals}
\alias{plot_tolerance_intervals}
\title{Plot 3: Credible vs. Tolerance Intervals for a Single Feature}
\usage{
plot_tolerance_intervals(post_matrix, feature_values, feature_name)
}
\description{
Plot 3: Credible vs. Tolerance Intervals for a Single Feature
}
