% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{predict.bart_samples}
\alias{predict.bart_samples}
\title{Predicts from a sampled BART model and returns the result as a matrix}
\usage{
\method{predict}{bart_samples}(bart, X, omega, num_samples)
}
\arguments{
\item{bart}{Object returned by the \code{bart} function}

\item{X}{Covariates for which outcome will be predicted}

\item{omega}{Basis vector to use in leaf node predictions}

\item{num_samples}{Number of retained samples in the original \code{bart} model}
}
\value{
Matrix with observations in rows and sampled BBART draws in columns
}
\description{
Predicts from a sampled BART model and returns the result as a matrix
}
\examples{
n <- 500
p <- 20
num_samples <- 50
num_burnin <- 10
num_trees <- 50
train_inds <- sample(1:n, round(n*0.8), replace = F)
test_inds <- (1:n)[!((1:n) \%in\% train_inds)]
X <- matrix(runif(n*p), ncol = p)
omega <- runif(n)
betas <- c(-5, -10, 10, 5)
y <- ifelse(
    (X[,1] >= 0) & (X[,1] < 0.25), betas[1] * omega,
    ifelse(
        (X[,1] >= 0.25) & (X[,1] < 0.5), betas[2] * omega,
        ifelse(
            (X[,1] >= 0.5) & (X[,1] < 0.75), betas[3] * omega, betas[4] * omega
        )
    )
) + rnorm(n, 0, 1)
nu <- 0.5
lambda <- 2.
stochtree_samples <- bart(ytrain, Xtrain, omegatrain, num_samples, num_burnin, num_trees, nu, lambda, random_seed = random_seed)
stochtree_predictions <- predict(stochtree_samples, Xtest, omegatest, num_samples)
}
