% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisel.R
\name{predict.bcfchiseler}
\alias{predict.bcfchiseler}
\title{Predict method for \code{bcfchiseler} objects}
\usage{
\method{predict}{bcfchiseler}(object, X_test, ...)
}
\arguments{
\item{object}{An object of class \code{bcfchiseler}.}

\item{X_test}{Matrix of covariates for test set.}

\item{...}{Additional arguments.}
}
\value{
A list containing posterior probabilities (\code{prob_smoothed}) and linear predictor draws (\code{boundary_linear_preds}).
}
\description{
Predict method for \code{bcfchiseler} objects
}
