% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict_forest}
\alias{predict_forest}
\title{Predict from a container of forest samples}
\usage{
predict_forest(forest_samples, dataset)
}
\arguments{
\item{forest_samples}{Pointer to a ForestContainer object}

\item{dataset}{Pointer to a ForestDataset object}
}
\value{
Matrix of predictions
}
\description{
Predict from a container of forest samples
}
