% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict_forest_raw}
\alias{predict_forest_raw}
\title{Extract raw leaf values from a container of forest samples}
\usage{
predict_forest_raw(forest_samples, dataset)
}
\arguments{
\item{forest_samples}{Pointer to a ForestContainer object}

\item{dataset}{Pointer to a ForestDataset object}
}
\value{
If leaves are univariates, matrix of predictions with
nrow equal to the number of observations in dataset and ncol
equal to the number of samples drawn in forest_samples.
if leaves are multivariate, an array of predictions with
dimensions equal to (1) the number of observations in
dataset, (2) the number of dimensions of the leaves in
forest_samples, and (3) the number of samples drawn in
forest_samples.
}
\description{
Extract raw leaf values from a container of forest samples
}
