% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_linear_probit.R
\name{predict_linear_bcf}
\alias{predict_linear_bcf}
\title{Predict from a sampled Linear Probit BCF model on new data}
\usage{
predict_linear_bcf(
  object,
  X,
  Z,
  propensity = NULL,
  rfx_group_ids = NULL,
  rfx_basis = NULL,
  ...
)
}
\arguments{
\item{object}{Object of type \code{bcfmodel} returned by \code{bcf_linear_probit}.}

\item{X}{Covariates used to determine tree leaf predictions for each observation. Must be passed as a matrix or dataframe.}

\item{Z}{Treatments used for prediction.}

\item{propensity}{(Optional) Propensities used for prediction.}

\item{rfx_group_ids}{(Optional) Test set group labels used for an additive random effects model.}

\item{rfx_basis}{(Optional) Test set basis for "random-slope" regression in additive random effects model.}
}
\value{
List of \code{nrow(X)} by \code{object$num_samples} matrices.
}
\description{
Predict from a sampled Linear Probit BCF model on new data
}
