% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessBartParams}
\alias{preprocessBartParams}
\title{Preprocess BART parameter list. Override defaults with any provided parameters.}
\usage{
preprocessBartParams(
  general_params,
  mean_forest_params,
  variance_forest_params
)
}
\arguments{
\item{general_params}{List of any non-forest-specific parameters}

\item{mean_forest_params}{List of any mean forest parameters}

\item{variance_forest_params}{List of any variance forest parameters}
}
\value{
Parameter list with defaults overriden by values supplied in parameter lists
}
\description{
Preprocess BART parameter list. Override defaults with any provided parameters.
}
