% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessBcfParams}
\alias{preprocessBcfParams}
\title{Preprocess BCF parameter list. Override defaults with any provided parameters.}
\usage{
preprocessBcfParams(params)
}
\arguments{
\item{general_params}{List of any non-forest-specific parameters}

\item{mu_forest_params}{List of any mu forest parameters}

\item{tau_forest_params}{List of any tau forest parameters}

\item{variance_forest_params}{List of any variance forest parameters}
}
\value{
Parameter list with defaults overriden by values supplied in parameter lists
}
\description{
Preprocess BCF parameter list. Override defaults with any provided parameters.
}
