% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessParams}
\alias{preprocessParams}
\title{Preprocess a parameter list, overriding defaults with any provided parameters.}
\usage{
preprocessParams(user_params, default_params)
}
\arguments{
\item{user_params}{User-supplied overrides to \code{default_params}.}

\item{default_params}{List of parameters with default values set.}
}
\value{
Parameter list with defaults overriden by values supplied in \code{user_params}
}
\description{
Preprocess a parameter list, overriding defaults with any provided parameters.
}
