% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessPredictionMatrix}
\alias{preprocessPredictionMatrix}
\title{Preprocess a matrix of covariate values, assuming all columns are numeric.}
\usage{
preprocessPredictionMatrix(input_matrix, metadata)
}
\arguments{
\item{metadata}{List containing information on variables, including train set
categories for categorical variables}

\item{input_df}{Covariate matrix.}
}
\value{
Preprocessed data with categorical variables appropriately preprocessed
}
\description{
Preprocess a matrix of covariate values, assuming all columns are numeric.
}
\examples{
cov_df <- data.frame(x1 = 1:5, x2 = 5:1, x3 = 6:10)
metadata <- list(num_ordered_cat_vars = 0, num_unordered_cat_vars = 0, num_numeric_vars = 3)
X_preprocessed <- preprocessPredictionDataFrame(cov_df, metadata)
}
