% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessTrainMatrix}
\alias{preprocessTrainMatrix}
\title{Preprocess a matrix of covariate values, assuming all columns are numeric.
Returns a list including a matrix of preprocessed covariate values and associated tracking.}
\usage{
preprocessTrainMatrix(input_matrix)
}
\arguments{
\item{input_matrix}{Covariate matrix.}
}
\value{
List with preprocessed (unmodified) data and details on the number of each type
of variable, unique categories associated with categorical variables, and the
vector of feature types needed for calls to BART and BCF.
}
\description{
Preprocess a matrix of covariate values, assuming all columns are numeric.
Returns a list including a matrix of preprocessed covariate values and associated tracking.
}
\examples{
cov_mat <- matrix(1:12, ncol = 3)
preprocess_list <- preprocessTrainMatrix(cov_mat)
X <- preprocess_list$X
}
