% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{print.bartmodel}
\alias{print.bartmodel}
\title{Print a summary of the BART model}
\usage{
\method{print}{bartmodel}(bart_model, ...)
}
\arguments{
\item{bart_model}{The BART model object}

\item{...}{Additional arguments}
}
\value{
BART model object unchanged after printing summary
}
\description{
Prints a summary of the BART model, including the model terms and their specifications.
}
