% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{print.bcfmodel}
\alias{print.bcfmodel}
\title{Print a summary of the BCF model}
\usage{
\method{print}{bcfmodel}(bcf_model, ...)
}
\arguments{
\item{...}{Additional arguments (currently unused)}

\item{bart_model}{The BCF model object}
}
\value{
BCF model object unchanged after printing summary
}
\description{
Prints a summary of the BCF model, including the model terms and their specifications.
}
