% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{print.ForestSamples}
\alias{print.ForestSamples}
\title{Summarize a ForestSamples object}
\usage{
\method{print}{ForestSamples}(x, ...)
}
\arguments{
\item{x}{ForestSamples object}

\item{...}{Additional arguments}
}
\value{
ForestSamples object unchanged after printing summary
}
\description{
Prints a summary of the ForestSamples object, including number of forests and the underlying model of each forest.
}
