% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{print.RandomEffectSamples}
\alias{print.RandomEffectSamples}
\title{Summarize a RandomEffectSamples object}
\usage{
\method{print}{RandomEffectSamples}(x, ...)
}
\arguments{
\item{x}{RandomEffectSamples object}

\item{...}{Additional arguments}
}
\value{
RandomEffectSamples object unchanged after printing summary
}
\description{
Prints a summary of the RandomEffectSamples object, including number of forests and the underlying model of each forest.
}
