% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{random_number_generator}
\alias{random_number_generator}
\title{Create a C++ random number generator and store its pointer (for replicability across sampling function calls)}
\usage{
random_number_generator(random_seed)
}
\arguments{
\item{random_seed}{Seed for the RNG}
}
\value{
External pointer to std::mt19937 object
}
\description{
Create a C++ random number generator and store its pointer (for replicability across sampling function calls)
}
\examples{
rng_ptr <- random_number_generator(101)
}
