% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{resetActiveForest}
\alias{resetActiveForest}
\title{Reset an active forest, either from a specific forest in a \code{ForestContainer}
or to an ensemble of single-node (i.e. root) trees}
\usage{
resetActiveForest(active_forest, forest_samples = NULL, forest_num = NULL)
}
\arguments{
\item{active_forest}{Current active forest}

\item{forest_samples}{(Optional) Container of forest samples from which to re-initialize active forest. If not provided, active forest will be reset to an ensemble of single-node (i.e. root) trees.}

\item{forest_num}{(Optional) Index of forest samples from which to initialize active forest. If not provided, active forest will be reset to an ensemble of single-node (i.e. root) trees.}
}
\description{
Reset an active forest, either from a specific forest in a \code{ForestContainer}
or to an ensemble of single-node (i.e. root) trees
}
