% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{resetForestModel}
\alias{resetForestModel}
\title{Re-initialize a forest model (tracking data structures) from a specific forest in a \code{ForestContainer}}
\usage{
resetForestModel(forest_model, dataset, forest_samples, forest_num)
}
\arguments{
\item{forest_model}{Forest model with tracking data structures}

\item{dataset}{Training dataset object}

\item{forest_samples}{Container of forest samples from which to re-initialize active forest}

\item{forest_num}{Index of forest samples from which to initialize active forest}
}
\value{
\code{Forest} object
}
\description{
Re-initialize a forest model (tracking data structures) from a specific forest in a \code{ForestContainer}
}
