% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{resetRandomEffectsModel}
\alias{resetRandomEffectsModel}
\title{Reset a \code{RandomEffectsModel} object based on the parameters indexed by \code{sample_num} in a \code{RandomEffectsSamples} object}
\usage{
resetRandomEffectsModel(rfx_model, rfx_samples, sample_num, sigma_alpha_init)
}
\arguments{
\item{rfx_samples}{}

\item{sample_num}{}

\item{rfx_tracker}{}
}
\description{
Reset a \code{RandomEffectsModel} object based on the parameters indexed by \code{sample_num} in a \code{RandomEffectsSamples} object
}
