% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{resetRandomEffectsTracker}
\alias{resetRandomEffectsTracker}
\title{Reset a \code{RandomEffectsTracker} object based on the parameters indexed by \code{sample_num} in a \code{RandomEffectsSamples} object}
\usage{
resetRandomEffectsTracker(
  rfx_tracker,
  rfx_model,
  rfx_dataset,
  residual,
  rfx_samples,
  sample_num
)
}
\arguments{
\item{rfx_tracker}{}

\item{rfx_dataset}{}

\item{residual}{}

\item{rfx_samples}{}

\item{sample_num}{}
}
\description{
Reset a \code{RandomEffectsTracker} object based on the parameters indexed by \code{sample_num} in a \code{RandomEffectsSamples} object
}
