% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{rootResetActiveForest}
\alias{rootResetActiveForest}
\title{Reset an active forest to an ensemble of single-node (i.e. root) trees}
\usage{
rootResetActiveForest(active_forest)
}
\arguments{
\item{active_forest}{Current active forest}
}
\value{
\code{Forest} object
}
\description{
Reset an active forest to an ensemble of single-node (i.e. root) trees
}
