% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{rootResetRandomEffectsModel}
\alias{rootResetRandomEffectsModel}
\title{Reset a \code{RandomEffectsModel} object to its "default" state}
\usage{
rootResetRandomEffectsModel(
  rfx_model,
  alpha_init,
  xi_init,
  sigma_alpha_init,
  sigma_xi_init,
  sigma_xi_shape,
  sigma_xi_scale
)
}
\arguments{
\item{xi_init}{}
}
\description{
Reset a \code{RandomEffectsModel} object to its "default" state
}
