% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{rootResetRandomEffectsTracker}
\alias{rootResetRandomEffectsTracker}
\title{Reset a \code{RandomEffectsTracker} object to its "default" state}
\usage{
rootResetRandomEffectsTracker(rfx_tracker, rfx_model, rfx_dataset, residual)
}
\arguments{
\item{rfx_tracker}{}

\item{rfx_dataset}{}

\item{residual}{}
}
\description{
Reset a \code{RandomEffectsTracker} object to its "default" state
}
