% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_model_one_iteration}
\alias{sample_model_one_iteration}
\title{Sample one iteration of the model}
\usage{
sample_model_one_iteration(
  data,
  residual,
  forest_samples,
  tracker,
  split_prior,
  rng,
  feature_types,
  leaf_model_int,
  leaf_model_scale,
  global_scale,
  cutpoint_grid_size = 500,
  gfr = T
)
}
\arguments{
\item{data}{Dataset (covariates and optional basis and variance weights)}

\item{residual}{Outcome}

\item{forest_samples}{Container for forest samples}

\item{tracker}{Wrapper around convenience data structures used in sampling}

\item{split_prior}{Prior on tree splits}

\item{rng}{Pointer to C++ random number generator}

\item{feature_types}{Vector of integers mapping to feature types (0 = numeric, 1 = ordered categorical, 2 = unordered categorical)}

\item{leaf_model_int}{Integer indicating the type of leaf model (0 = constant, 1 = univariate regression, 2 = multivariate regression)}

\item{leaf_model_scale}{Value of leaf model scale parameter (passed as a matrix, handled as a double in C++ if the leaf model is univariate)}

\item{global_scale}{Value of global variance parameter}

\item{cutpoint_grid_size}{Maximum number of cutpoints to evaluate}

\item{gfr}{(Boolean) Whether or not the GFR sampler should be used (if false, MCMC is used)}
}
\description{
Sample one iteration of the model
}
