% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_sigma2_one_iteration}
\alias{sample_sigma2_one_iteration}
\title{Sample one iteration of the global variance model}
\usage{
sample_sigma2_one_iteration(residual, rng, nu, lambda)
}
\arguments{
\item{residual}{Pointer to a C++ ColumnVector object storing residual}

\item{rng}{Pointer to a C++ random number generator}

\item{nu}{Global variance shape parameter}

\item{lambda}{Constitutes the scale parameter for the global variance along with nu (i.e. scale is nu*lambda)}
}
\description{
Sample one iteration of the global variance model
}
