% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{savePreprocessorToJsonString}
\alias{savePreprocessorToJsonString}
\title{Convert the persistent aspects of a covariate preprocessor to (in-memory) JSON string}
\usage{
savePreprocessorToJsonString(object)
}
\arguments{
\item{object}{List containing information on variables, including train set
categories for categorical variables}
}
\value{
in-memory JSON string
}
\description{
Convert the persistent aspects of a covariate preprocessor to (in-memory) JSON string
}
\examples{
cov_mat <- matrix(1:12, ncol = 3)
preprocess_list <- preprocessTrainData(cov_mat)
preprocessor_json_string <- savePreprocessorToJsonString(preprocess_list$metadata)
}
