% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{shapley}
\alias{shapley}
\title{Shapley calculator for x_{i}}
\usage{
shapley(index, X, beta_post, beta_int_post)
}
\arguments{
\item{index:}{row index in X for which to compute Shapley values}

\item{X:}{centered covariate matrix (n x p)}

\item{beta_post:}{posterior samples of beta (chains x samples x p)}

\item{beta_int_post:}{posterior samples of interaction effects (chains x samples x p_int)}
}
\value{
The posterior samples of the shapley values, and a dataframe with summary statistics.
}
\description{
Shapley calculator.
}
