% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley.R
\name{shapley_lean}
\alias{shapley_lean}
\title{Highly Optimized & Vectorized Shapley Calculator for a Single Observation}
\usage{
shapley_lean(
  index,
  x_star,
  cov_X,
  beta_samples,
  beta_int_samples,
  ipairs,
  p_int,
  num_covariates,
  col_names
)
}
\description{
Highly Optimized & Vectorized Shapley Calculator for a Single Observation
}
