% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stochtree.params2str}
\alias{stochtree.params2str}
\title{Utility function from lightGBM's R package.
Converts R-style list of parameters to a string of
the form "param1=-1,param2=yes,..."}
\usage{
stochtree.params2str(params)
}
\arguments{
\item{params}{List of parameters as inputs to either the dataset
creation process or the model training process}
}
\value{
String containing parameters and their values
}
\description{
Renamed here to reflect the naming convention of this package, but
the code is the same.
}
\examples{
param_list <- list(param1 = -1, param2 = "yes")
param_string <- stochtree.params2str(param_list)
}
