% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{summary.bcfmodel}
\alias{summary.bcfmodel}
\title{Summarize BCF Model}
\usage{
\method{summary}{bcfmodel}(object, ...)
}
\arguments{
\item{object}{The BCF model object}

\item{...}{Additional arguments}
}
\value{
BCF model object unchanged after summarizing
}
\description{
Summarize the BCF with a description of the model that was fit and numeric summaries of any sampled quantities.
}
