% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{tree_prior}
\alias{tree_prior}
\title{Create a StochTree tree prior object}
\usage{
tree_prior(alpha, beta, min_samples_leaf)
}
\arguments{
\item{alpha}{Split probability at root}

\item{beta}{Decay parameter for split probability}

\item{min_samples_leaf}{Minimum number of samples in a valid leaf}
}
\value{
External pointer to tree prior object
}
\description{
Create a StochTree tree prior object
}
\examples{
tree_prior_ptr <- tree_prior(0.95, 1.25, 10)
}
