% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{UpliftForestDataset}
\alias{UpliftForestDataset}
\title{Dataset used to sample an uplift forest}
\description{
An uplift dataset consists of four matrices / vectors: covariates,
treatment indicator, bases, and variance weights.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpliftForestDataset-new}{\code{UpliftForestDataset$new()}}
\item \href{#method-UpliftForestDataset-update_basis}{\code{UpliftForestDataset$update_basis()}}
\item \href{#method-UpliftForestDataset-update_weights}{\code{UpliftForestDataset$update_weights()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpliftForestDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpliftForestDataset-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpliftForestDataset$new(
  covariates,
  treatment,
  basis = NULL,
  variance_weights = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpliftForestDataset-update_basis"></a>}}
\if{latex}{\out{\hypertarget{method-UpliftForestDataset-update_basis}{}}}
\subsection{Method \code{update_basis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpliftForestDataset$update_basis(basis)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpliftForestDataset-update_weights"></a>}}
\if{latex}{\out{\hypertarget{method-UpliftForestDataset-update_weights}{}}}
\subsection{Method \code{update_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpliftForestDataset$update_weights(weights)}\if{html}{\out{</div>}}
}

}
}
