% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbart.R
\name{xbart}
\alias{xbart}
\title{Samples from an XBART model using the provided data}
\usage{
xbart(model_matrix, params = list())
}
\arguments{
\item{model_matrix}{Matrix containing data (including outcome of interest and treatment if present)}

\item{params}{List of parameters of the form \code{list(param1=-1,...)}}
}
\value{
Object of class XBART
}
\description{
Samples from an XBART model using the provided data
}
\examples{
n <- 1000
p <- 5
X <- matrix(runif(n*p), ncol=p)
y <- X[,1] * 5 + X[,2] * 10 + rnorm(n, 0, 0.5)
model_matrix <- cbind(y, X)
param_list = list(label_column=0, num_trees=50, num_burnin=0, num_samples=20, min_samples_in_leaf=5)
result <- xbart(model_matrix, param_list)
}
