% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart}
\alias{bart}
\title{Samples from a BART model using the provided data}
\usage{
bart(
  y,
  X,
  omega,
  num_samples,
  num_burnin,
  num_trees,
  nu,
  lambda,
  random_seed = -1
)
}
\arguments{
\item{y}{Outcome of interest}

\item{X}{Covariates used in the tree model}

\item{omega}{Basis functions used for leaf node regression}

\item{num_samples}{Number of samples retained}

\item{num_burnin}{Number of "burn-in" (not retained) samples}

\item{num_trees}{Number of trees in the ensemble}

\item{nu}{Prior scale for global variance parameter (sigma^2)}

\item{lambda}{Prior shape for global variance parameter (sigma^2)}

\item{random_seed}{Seed for random number generator}
}
\value{
Object of class \code{bart_samples}
}
\description{
Samples from a BART model using the provided data
}
\examples{
n <- 500
p <- 20
num_samples <- 50
num_burnin <- 10
num_trees <- 50
train_inds <- sample(1:n, round(n*0.8), replace = F)
test_inds <- (1:n)[!((1:n) \%in\% train_inds)]
X <- matrix(runif(n*p), ncol = p)
omega <- runif(n)
betas <- c(-5, -10, 10, 5)
y <- ifelse(
    (X[,1] >= 0) & (X[,1] < 0.25), betas[1] * omega,
    ifelse(
        (X[,1] >= 0.25) & (X[,1] < 0.5), betas[2] * omega,
        ifelse(
            (X[,1] >= 0.5) & (X[,1] < 0.75), betas[3] * omega, betas[4] * omega
        )
    )
) + rnorm(n, 0, 1)
cutpoint_grid_size <- 20
nu <- 0.5
lambda <- 2.
a <- 1.
b <- 1.
result <- bart(ytrain, Xtrain, omegatrain, num_samples, num_burnin, num_trees, nu, lambda, random_seed = random_seed)
}
