% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_score_test.R
\name{bcf_restricted_score_test}
\alias{bcf_restricted_score_test}
\title{Fit a Restricted BCF Model (Constant Treatment Effect) and Compute Score Test P-Values}
\usage{
bcf_restricted_score_test(
  X_train,
  Z_train,
  y_train,
  propensity_train = NULL,
  rfx_group_ids_train = NULL,
  rfx_basis_train = NULL,
  num_gfr = 5,
  num_burnin = 100,
  num_mcmc = 500,
  general_params = list(),
  prognostic_forest_params = list(),
  variance_forest_params = list()
)
}
\arguments{
\item{X_train}{Covariates used to split trees in the ensemble.}

\item{Z_train}{Vector of treatment assignments.}

\item{y_train}{Outcome to be modeled by the ensemble.}

\item{propensity_train}{(Optional) Vector of propensity scores.}

\item{rfx_group_ids_train}{(Optional) Group labels used for an additive random effects model.}

\item{rfx_basis_train}{(Optional) Basis for "random-slope" regression.}

\item{num_gfr}{Number of "warm-start" iterations. Default: 5.}

\item{num_burnin}{Number of "burn-in" iterations. Default: 100.}

\item{num_mcmc}{Number of "retained" iterations. Default: 500.}

\item{general_params}{(Optional) A list of general model parameters.}

\item{prognostic_forest_params}{(Optional) A list of prognostic forest model parameters.}

\item{variance_forest_params}{(Optional) A list of variance forest model parameters.}
}
\value{
A list of class \code{bcfscoretest} containing posterior samples and score test P-values.
}
\description{
Fit a Restricted BCF Model (Constant Treatment Effect) and Compute Score Test P-Values
}
