% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{createCppJsonString}
\alias{createCppJsonString}
\title{Create a C++ Json object from a Json string}
\usage{
createCppJsonString(json_string)
}
\arguments{
\item{json_string}{JSON string dump}
}
\value{
\code{CppJson} object
}
\description{
Create a C++ Json object from a Json string
}
