% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{createCppJsonString}
\alias{createCppJsonString}
\title{Create a C++ Json object from a Json string}
\usage{
createCppJsonString(json_string)
}
\arguments{
\item{json_string}{JSON string dump}
}
\value{
\code{CppJson} object
}
\description{
Create a C++ Json object from a Json string
}
\examples{
example_vec <- runif(10)
example_json <- createCppJson()
example_json$add_vector("myvec", example_vec)
example_json_string <- example_json$return_json_string()
example_json_roundtrip <- createCppJsonString(example_json_string)
}
