% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{createForest}
\alias{createForest}
\title{Create a forest}
\usage{
createForest(
  num_trees,
  leaf_dimension = 1,
  is_leaf_constant = FALSE,
  is_exponentiated = FALSE
)
}
\arguments{
\item{num_trees}{Number of trees in the forest}

\item{leaf_dimension}{Dimensionality of the outcome model}

\item{is_leaf_constant}{Whether leaf is constant}

\item{is_exponentiated}{Whether forest predictions should be exponentiated before being returned}
}
\value{
\code{Forest} object
}
\description{
Create a forest
}
\examples{
num_trees <- 100
leaf_dimension <- 2
is_leaf_constant <- FALSE
is_exponentiated <- FALSE
forest <- createForest(num_trees, leaf_dimension, is_leaf_constant, is_exponentiated)
}
