% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createForestModel}
\alias{createForestModel}
\title{Create a forest model object}
\usage{
createForestModel(
  forest_dataset,
  feature_types,
  num_trees,
  n,
  alpha,
  beta,
  min_samples_leaf,
  max_depth
)
}
\arguments{
\item{forest_dataset}{\code{ForestDataset} object, used to initialize forest sampling data structures}

\item{feature_types}{Feature types (integers where 0 = numeric, 1 = ordered categorical, 2 = unordered categorical)}

\item{num_trees}{Number of trees in the forest being sampled}

\item{n}{Number of observations in \code{forest_dataset}}

\item{alpha}{Root node split probability in tree prior}

\item{beta}{Depth prior penalty in tree prior}

\item{min_samples_leaf}{Minimum number of samples in a tree leaf}

\item{max_depth}{Maximum depth of any tree in the ensemble in the mean model. Setting to \code{-1} does not enforce any depth limits on trees.}
}
\value{
\code{ForestModel} object
}
\description{
Create a forest model object
}
