% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{createGlobalModelConfig}
\alias{createGlobalModelConfig}
\title{Create a global model config object}
\usage{
createGlobalModelConfig(global_error_variance = 1)
}
\arguments{
\item{global_error_variance}{Global error variance parameter (default: \code{1.0})}
}
\value{
GlobalModelConfig object
}
\description{
Create a global model config object
}
\examples{
config <- createGlobalModelConfig(global_error_variance = 100)
}
