% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{createOutcome}
\alias{createOutcome}
\title{Create an outcome object}
\usage{
createOutcome(outcome)
}
\arguments{
\item{outcome}{Vector of outcome values}
}
\value{
\code{Outcome} object
}
\description{
Create an outcome object
}
