% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{createRandomEffectSamples}
\alias{createRandomEffectSamples}
\title{Create a \code{RandomEffectSamples} object}
\usage{
createRandomEffectSamples(num_components, num_groups, random_effects_tracker)
}
\arguments{
\item{num_components}{Number of "components" or bases defining the random effects regression}

\item{num_groups}{Number of random effects groups}

\item{random_effects_tracker}{Object of type \code{RandomEffectsTracker}}
}
\value{
\code{RandomEffectSamples} object
}
\description{
Create a \code{RandomEffectSamples} object
}
