% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{createRandomEffectsTracker}
\alias{createRandomEffectsTracker}
\title{Create a \code{RandomEffectsTracker} object}
\usage{
createRandomEffectsTracker(rfx_group_indices)
}
\arguments{
\item{rfx_group_indices}{Integer indices indicating groups used to define random effects}
}
\value{
\code{RandomEffectsTracker} object
}
\description{
Create a \code{RandomEffectsTracker} object
}
