% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{ForestSamples}
\alias{ForestSamples}
\title{Class that stores draws from an random ensemble of decision trees}
\description{
Wrapper around a C++ container of tree ensembles
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{forest_container_ptr}}{External pointer to a C++ ForestContainer class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ForestSamples-new}{\code{ForestSamples$new()}}
\item \href{#method-ForestSamples-load_from_json}{\code{ForestSamples$load_from_json()}}
\item \href{#method-ForestSamples-predict}{\code{ForestSamples$predict()}}
\item \href{#method-ForestSamples-predict_raw}{\code{ForestSamples$predict_raw()}}
\item \href{#method-ForestSamples-predict_raw_single_forest}{\code{ForestSamples$predict_raw_single_forest()}}
\item \href{#method-ForestSamples-set_root_leaves}{\code{ForestSamples$set_root_leaves()}}
\item \href{#method-ForestSamples-update_residual}{\code{ForestSamples$update_residual()}}
\item \href{#method-ForestSamples-save_json}{\code{ForestSamples$save_json()}}
\item \href{#method-ForestSamples-load_json}{\code{ForestSamples$load_json()}}
\item \href{#method-ForestSamples-num_samples}{\code{ForestSamples$num_samples()}}
\item \href{#method-ForestSamples-num_trees}{\code{ForestSamples$num_trees()}}
\item \href{#method-ForestSamples-output_dimension}{\code{ForestSamples$output_dimension()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-new"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-new}{}}}
\subsection{Method \code{new()}}{
Create a new ForestContainer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$new(num_trees, output_dimension = 1, is_leaf_constant = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_trees}}{Number of trees}

\item{\code{output_dimension}}{Dimensionality of the outcome model}

\item{\code{is_leaf_constant}}{Whether leaf is constant}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ForestContainer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-load_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-load_from_json}{}}}
\subsection{Method \code{load_from_json()}}{
Create a new ForestContainer object from a json object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$load_from_json(json_object, json_forest_label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_object}}{Object of class \code{CppJson}}

\item{\code{json_forest_label}}{Label referring to a particular forest (i.e. "forest_0") in the overall json hierarchy}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ForestContainer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-predict"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-predict}{}}}
\subsection{Method \code{predict()}}{
Predict every tree ensemble on every sample in \code{forest_dataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$predict(forest_dataset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forest_dataset}}{\code{ForestDataset} R class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matrix of predictions with as many rows as in forest_dataset
and as many columns as samples in the \code{ForestContainer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-predict_raw"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-predict_raw}{}}}
\subsection{Method \code{predict_raw()}}{
Predict "raw" leaf values (without being multiplied by basis) for every tree ensemble on every sample in \code{forest_dataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$predict_raw(forest_dataset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forest_dataset}}{\code{ForestDataset} R class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array of predictions for each observation in \code{forest_dataset} and
each sample in the \code{ForestSamples} class with each prediction having the
dimensionality of the forests' leaf model. In the case of a constant leaf model
or univariate leaf regression, this array is two-dimensional (number of observations,
number of forest samples). In the case of a multivariate leaf regression,
this array is three-dimension (number of observations, leaf model dimension,
number of samples).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-predict_raw_single_forest"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-predict_raw_single_forest}{}}}
\subsection{Method \code{predict_raw_single_forest()}}{
Predict "raw" leaf values (without being multiplied by basis) for a specific forest on every sample in \code{forest_dataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$predict_raw_single_forest(forest_dataset, forest_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forest_dataset}}{\code{ForestDataset} R class}

\item{\code{forest_num}}{Index of the forest sample within the container}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matrix of predictions with as many rows as in forest_dataset
and as many columns as samples in the \code{ForestContainer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-set_root_leaves"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-set_root_leaves}{}}}
\subsection{Method \code{set_root_leaves()}}{
Set a constant predicted value for every tree in the ensemble.
Stops program if any tree is more than a root node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$set_root_leaves(forest_num, leaf_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forest_num}}{Index of the forest sample within the container.}

\item{\code{leaf_value}}{Constant leaf value(s) to be fixed for each tree in the ensemble indexed by \code{forest_num}. Can be either a single number or a vector, depending on the forest's leaf dimension.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-update_residual"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-update_residual}{}}}
\subsection{Method \code{update_residual()}}{
Updates residual based on the predictions of a forest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$update_residual(
  dataset,
  outcome,
  forest_model,
  requires_basis,
  forest_num,
  add
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{\code{ForestDataset} object storing the covariates and bases for a given forest}

\item{\code{outcome}}{\code{Outcome} object storing the residuals to be updated based on forest predictions}

\item{\code{forest_model}}{\code{ForestModel} object storing tracking structures used in training / sampling}

\item{\code{requires_basis}}{Whether or not a forest requires a basis for prediction}

\item{\code{forest_num}}{Index of forest used to update residuals}

\item{\code{add}}{Whether forest predictions should be added to or subtracted from residuals}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-save_json"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-save_json}{}}}
\subsection{Method \code{save_json()}}{
Store the trees and metadata of \code{ForestDataset} class in a json file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$save_json(json_filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_filename}}{Name of output json file (must end in ".json")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-load_json"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-load_json}{}}}
\subsection{Method \code{load_json()}}{
Load trees and metadata for an ensemble from a json file. Note that
any trees and metadata already present in \code{ForestDataset} class will
be overwritten.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$load_json(json_filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json_filename}}{Name of model input json file (must end in ".json")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-num_samples"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-num_samples}{}}}
\subsection{Method \code{num_samples()}}{
Return number of samples in a \code{ForestContainer} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$num_samples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Sample count
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-num_trees"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-num_trees}{}}}
\subsection{Method \code{num_trees()}}{
Return number of trees in each ensemble of a \code{ForestContainer} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$num_trees()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Tree count
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ForestSamples-output_dimension"></a>}}
\if{latex}{\out{\hypertarget{method-ForestSamples-output_dimension}{}}}
\subsection{Method \code{output_dimension()}}{
Return output dimension of trees in a \code{ForestContainer} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ForestSamples$output_dimension()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Leaf node parameter size
}
}
}
