% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_sampler.R
\name{horseshoe_sampler}
\alias{horseshoe_sampler}
\title{Fit a heteroskedastic Bayesian linear regression model with a horseshoe prior}
\usage{
horseshoe_sampler(
  X,
  y,
  z,
  num_iterations = 2000,
  burn_in = 1000,
  mu_model = list(gibbs = TRUE, global_shrink = TRUE, unlink = FALSE, p_int = 0),
  tau_model = list(gibbs = TRUE, global_shrink = TRUE, unlink = FALSE, p_int = 0),
  propensity_separate = FALSE,
  alpha_prior_sd = 10,
  initial_values = NULL,
  standardize_cov = F,
  interaction_rule = "continuous_or_binary"
)
}
\arguments{
\item{X}{A numeric matrix or data frame of covariates.}

\item{y}{A numeric vector of outcomes.}

\item{z}{A numeric vector of treatment assignments (e.g., 0 or 1).}

\item{num_iterations}{Total number of MCMC iterations.}

\item{burn_in}{Number of burn-in iterations to discard.}

\item{mu_model}{A list of options for the mu(x) model component.}

\item{tau_model}{A list of options for the tau(x) model component.}

\item{propensity_separate}{(logical) Whether to separately model the propensity score.}

\item{alpha_prior_sd}{(double) The standard deviation of the prior on the intercept terms.}

\item{initial_values}{A list of starting values for the sampler. If NULL (default), data driven initiliaziion will be made.}

\item{standardize_cov}{test}

\item{interaction_rule}{test}
}
\value{
A list object of class \code{stochtree_hs} containing the posterior samples for all model parameters.
}
\description{
This function fits a model of the form y = mu(x) + z * tau(x) + epsilon,
where mu(x) is the prognostic effect and tau(x) is the treatment effect.
Both mu(x) and tau(x) are linear models with coefficients given a horseshoe
prior. This allows for flexible modeling of treatment effect heterogeneity.
}
