% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{loadForestContainerJson}
\alias{loadForestContainerJson}
\title{Load a container of forest samples from json}
\usage{
loadForestContainerJson(json_object, json_forest_label)
}
\arguments{
\item{json_object}{Object of class \code{CppJson}}

\item{json_forest_label}{Label referring to a particular forest (i.e. "forest_0") in the overall json hierarchy}
}
\value{
\code{ForestSamples} object
}
\description{
Load a container of forest samples from json
}
