% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{loadForestContainerJson}
\alias{loadForestContainerJson}
\title{Load a container of forest samples from json}
\usage{
loadForestContainerJson(json_object, json_forest_label)
}
\arguments{
\item{json_object}{Object of class \code{CppJson}}

\item{json_forest_label}{Label referring to a particular forest (i.e. "forest_0") in the overall json hierarchy}
}
\value{
\code{ForestSamples} object
}
\description{
Load a container of forest samples from json
}
\examples{
X <- matrix(runif(10*100), ncol = 10)
y <- -5 + 10*(X[,1] > 0.5) + rnorm(100)
bart_model <- bart(X, y, num_gfr=0, num_mcmc=10)
bart_json <- saveBARTModelToJson(bart_model)
mean_forest <- loadForestContainerJson(bart_json, "forest_0")
}
