% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{loadRandomEffectSamplesJson}
\alias{loadRandomEffectSamplesJson}
\title{Load a container of random effect samples from json}
\usage{
loadRandomEffectSamplesJson(json_object, json_rfx_num)
}
\arguments{
\item{json_object}{Object of class \code{CppJson}}

\item{json_rfx_num}{Integer index indicating the position of the random effects term to be unpacked}
}
\value{
\code{RandomEffectSamples} object
}
\description{
Load a container of random effect samples from json
}
