% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Outcome}
\alias{Outcome}
\title{Outcome / partial residual used to sample an additive model.}
\description{
The outcome class is wrapper around a vector of (mutable)
outcomes for ML tasks (supervised learning, causal inference).
When an additive tree ensemble is sampled, the outcome used to
sample a specific model term is the "partial residual" consisting
of the outcome minus the predictions of every other model term
(trees, group random effects, etc...).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_ptr}}{External pointer to a C++ Outcome class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Outcome-new}{\code{Outcome$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outcome-new"></a>}}
\if{latex}{\out{\hypertarget{method-Outcome-new}{}}}
\subsection{Method \code{new()}}{
Create a new Outcome object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outcome$new(outcome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outcome}}{Vector of outcome values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Outcome} object.
}
}
}
