% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessBartParams}
\alias{preprocessBartParams}
\title{Preprocess BART parameter list. Override defaults with any provided parameters.}
\usage{
preprocessBartParams(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Parameter list with defaults overriden by values supplied in \code{params}
}
\description{
Preprocess BART parameter list. Override defaults with any provided parameters.
}
