% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessBcfParams}
\alias{preprocessBcfParams}
\title{Preprocess BCF parameter list. Override defaults with any provided parameters.}
\usage{
preprocessBcfParams(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Parameter list with defaults overriden by values supplied in \code{params}
}
\description{
Preprocess BCF parameter list. Override defaults with any provided parameters.
}
