% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessPredictionMatrix}
\alias{preprocessPredictionMatrix}
\title{Preprocess a matrix of covariate values, assuming all columns are numeric.}
\usage{
preprocessPredictionMatrix(input_matrix, metadata)
}
\arguments{
\item{input_matrix}{Covariate matrix.}

\item{metadata}{List containing information on variables, including train set
categories for categorical variables}
}
\value{
Preprocessed data with categorical variables appropriately preprocessed
}
\description{
Preprocess a matrix of covariate values, assuming all columns are numeric.
}
\examples{
cov_mat <- matrix(c(1:5, 5:1, 6:10), ncol = 3)
metadata <- list(num_ordered_cat_vars = 0, num_unordered_cat_vars = 0, 
                 num_numeric_vars = 3, numeric_vars = c("x1", "x2", "x3"))
X_preprocessed <- preprocessPredictionMatrix(cov_mat, metadata)
}
