% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{print.bartmodel}
\alias{print.bartmodel}
\title{Summarize a BART Model}
\usage{
\method{print}{bartmodel}(x, ...)
}
\arguments{
\item{x}{The BART model object}

\item{...}{Additional arguments}
}
\value{
BART model object unchanged after printing summary
}
\description{
Prints a summary of the BART model, including the model terms and their specifications.
}
