% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{print.bcfmodel}
\alias{print.bcfmodel}
\title{Print Summary of BCF Model}
\usage{
\method{print}{bcfmodel}(x, ...)
}
\arguments{
\item{x}{The BCF model object}

\item{...}{Additional arguments (currently unused)}
}
\value{
BCF model object unchanged after printing summary
}
\description{
Prints a summary of the BCF model, including the model terms and their specifications.
}
