% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{resetActiveForest}
\alias{resetActiveForest}
\title{Re-initialize an active forest from a specific forest in a \code{ForestContainer}}
\usage{
resetActiveForest(active_forest, forest_samples, forest_num)
}
\arguments{
\item{active_forest}{Current active forest}

\item{forest_samples}{Container of forest samples from which to re-initialize active forest}

\item{forest_num}{Index of forest samples from which to initialize active forest}
}
\description{
Re-initialize an active forest from a specific forest in a \code{ForestContainer}
}
