% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{resetForestModel}
\alias{resetForestModel}
\title{Re-initialize a forest model (tracking data structures) from a specific forest in a \code{ForestContainer}}
\usage{
resetForestModel(forest_model, forest, dataset, residual, is_mean_model)
}
\arguments{
\item{forest_model}{Forest model with tracking data structures}

\item{forest}{Forest from which to re-initialize forest model}

\item{dataset}{Training dataset object}

\item{residual}{Residual which will also be updated}

\item{is_mean_model}{Whether the model being updated is a conditional mean model}
}
\description{
Re-initialize a forest model (tracking data structures) from a specific forest in a \code{ForestContainer}
}
