% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{resetRandomEffectsTracker}
\alias{resetRandomEffectsTracker}
\title{Reset a \code{RandomEffectsTracker} object based on the parameters indexed by \code{sample_num} in a \code{RandomEffectsSamples} object}
\usage{
resetRandomEffectsTracker(
  rfx_tracker,
  rfx_model,
  rfx_dataset,
  residual,
  rfx_samples
)
}
\arguments{
\item{rfx_tracker}{Object of type \code{RandomEffectsTracker}.}

\item{rfx_model}{Object of type \code{RandomEffectsModel}.}

\item{rfx_dataset}{Object of type \code{RandomEffectsDataset}.}

\item{residual}{Object of type \code{Outcome}.}

\item{rfx_samples}{Object of type \code{RandomEffectSamples}.}
}
\description{
Reset a \code{RandomEffectsTracker} object based on the parameters indexed by \code{sample_num} in a \code{RandomEffectsSamples} object
}
