% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{rootResetRandomEffectsModel}
\alias{rootResetRandomEffectsModel}
\title{Reset a \code{RandomEffectsModel} object to its "default" state}
\usage{
rootResetRandomEffectsModel(
  rfx_model,
  alpha_init,
  xi_init,
  sigma_alpha_init,
  sigma_xi_init,
  sigma_xi_shape,
  sigma_xi_scale
)
}
\arguments{
\item{rfx_model}{Object of type \code{RandomEffectsModel}.}

\item{alpha_init}{Initial value of the "working parameter".}

\item{xi_init}{Initial value of the "group parameters".}

\item{sigma_alpha_init}{Initial value of the "working parameter" scale parameter.}

\item{sigma_xi_init}{Initial value of the "group parameters" scale parameter.}

\item{sigma_xi_shape}{Shape parameter for the inverse gamma variance model on the group parameters.}

\item{sigma_xi_scale}{Scale parameter for the inverse gamma variance model on the group parameters.}
}
\description{
Reset a \code{RandomEffectsModel} object to its "default" state
}
