% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effects.R
\name{rootResetRandomEffectsTracker}
\alias{rootResetRandomEffectsTracker}
\title{Reset a \code{RandomEffectsTracker} object to its "default" state}
\usage{
rootResetRandomEffectsTracker(rfx_tracker, rfx_model, rfx_dataset, residual)
}
\arguments{
\item{rfx_tracker}{Object of type \code{RandomEffectsTracker}.}

\item{rfx_model}{Object of type \code{RandomEffectsModel}.}

\item{rfx_dataset}{Object of type \code{RandomEffectsDataset}.}

\item{residual}{Object of type \code{Outcome}.}
}
\description{
Reset a \code{RandomEffectsTracker} object to its "default" state
}
