% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{sample_sigma2_one_iteration}
\alias{sample_sigma2_one_iteration}
\title{Sample one iteration of the (inverse gamma) global variance model}
\usage{
sample_sigma2_one_iteration(residual, dataset, rng, a, b)
}
\arguments{
\item{residual}{Outcome class}

\item{dataset}{ForestDataset class}

\item{rng}{C++ random number generator}

\item{a}{Global variance shape parameter}

\item{b}{Global variance scale parameter}
}
\description{
Sample one iteration of the (inverse gamma) global variance model
}
