% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{sample_tau_one_iteration}
\alias{sample_tau_one_iteration}
\title{Sample one iteration of the leaf parameter variance model (only for univariate basis and constant leaf!)}
\usage{
sample_tau_one_iteration(forest, rng, a, b)
}
\arguments{
\item{forest}{C++ forest}

\item{rng}{C++ random number generator}

\item{a}{Leaf variance shape parameter}

\item{b}{Leaf variance scale parameter}
}
\description{
Sample one iteration of the leaf parameter variance model (only for univariate basis and constant leaf!)
}
