% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochtree-package.R
\docType{package}
\name{stochtree-package}
\alias{stochtree}
\alias{stochtree-package}
\title{stochtree: Stochastic tree Ensembles (XBART and BART) for Supervised Learning and Causal Inference}
\description{
Flexible stochastic tree ensemble software. Robust implementations of Bayesian Additive Regression Trees (Chipman, George, McCulloch (2010) \doi{10.1214/09-AOAS285}) for supervised learning and (Bayesian Causal Forests (BCF) Hahn, Murray, Carvalho (2020) \doi{10.1214/19-BA1195}) for causal inference. Enables model serialization and parallel sampling and provides a low-level interface for custom stochastic forest samplers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stochtree.ai}
}

}
\author{
\strong{Maintainer}: Drew Herren \email{drewherrenopensource@gmail.com} (\href{https://orcid.org/0000-0003-4109-6611}{ORCID})

Authors:
\itemize{
  \item Richard Hahn
  \item Jared Murray
  \item Carlos Carvalho
  \item Jingyu He
}

Other contributors:
\itemize{
  \item stochtree contributors [copyright holder]
}

}
