% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochtree-package.R
\docType{package}
\name{stochtree-package}
\alias{stochtree}
\alias{stochtree-package}
\title{stochtree: Stochastic tree ensembles (XBART and BART) for supervised learning and causal inference}
\description{
Stochastic tree ensembles (XBART and BART) for supervised learning and causal inference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stochastictree.github.io/stochtree-r/}
}

}
\author{
\strong{Maintainer}: Drew Herren \email{drewherrenopensource@gmail.com} (\href{https://orcid.org/0000-0003-4109-6611}{ORCID})

Authors:
\itemize{
  \item Richard Hahn
  \item Jared Murray
  \item Carlos Carvalho
  \item Jingyu He
}

}
