% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{summary.bartmodel}
\alias{summary.bartmodel}
\title{Summarize the BART model fit and sampled terms.}
\usage{
\method{summary}{bartmodel}(object, ...)
}
\arguments{
\item{object}{The BART model object}

\item{...}{Additional arguments}
}
\value{
BART model object unchanged after summarizing
}
\description{
Summarize the BART with a description of the model that was fit and numeric summaries of any sampled quantities.
}
