% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbart.R
\name{xbart}
\alias{xbart}
\title{Samples from an XBART model using the provided data}
\usage{
xbart(
  y,
  X,
  omega,
  rfx_basis,
  rfx_groups,
  num_samples,
  num_burnin,
  num_trees,
  nu,
  lambda,
  a_leaf,
  b_leaf,
  a_rfx,
  b_rfx,
  cutpoint_grid_size,
  random_seed = -1
)
}
\arguments{
\item{y}{Outcome of interest}

\item{X}{Covariates used in the tree model}

\item{omega}{Basis functions used for leaf node regression}

\item{rfx_basis}{Basis function for group random effects}

\item{rfx_groups}{Group indices for random effects}

\item{num_samples}{Number of samples retained}

\item{num_burnin}{Number of "burn-in" (not retained) samples}

\item{num_trees}{Number of trees in the ensemble}

\item{nu}{Prior scale for global variance parameter (sigma^2)}

\item{lambda}{Prior shape for global variance parameter (sigma^2)}

\item{a_leaf}{Prior scale for leaf node variance parameter (tau)}

\item{b_leaf}{Prior shape for leaf node variance parameter (tau)}

\item{a_rfx}{Prior scale for group random effects}

\item{b_rfx}{Prior shape for group random effects}

\item{cutpoint_grid_size}{Maximum number of cutpoints to consider at each split}

\item{random_seed}{Seed for random number generator}
}
\value{
Object of class \code{xbart_samples}
}
\description{
Samples from an XBART model using the provided data
}
\examples{
n <- 500
p <- 20
num_samples <- 50
num_burnin <- 10
num_trees <- 50
train_inds <- sample(1:n, round(n*0.8), replace = F)
test_inds <- (1:n)[!((1:n) \%in\% train_inds)]
X <- matrix(runif(n*p), ncol = p)
rfx_labels <- as.integer(rep(c(1,2), times = n/2))
rfx_basis <- rep(1, times = n)
omega <- runif(n)
betas <- c(-5, -10, 10, 5)
f_x_omega <- ifelse(
    (X[,1] >= 0) & (X[,1] < 0.25), betas[1] * omega,
    ifelse(
        (X[,1] >= 0.25) & (X[,1] < 0.5), betas[2] * omega,
        ifelse(
            (X[,1] >= 0.5) & (X[,1] < 0.75), betas[3] * omega, betas[4] * omega
        )
    )
)
rfx <- ifelse(rfx_labels == 1, -1, 5)
y <- f_x_omega + rfx + rnorm(n, 0, 1)
cutpoint_grid_size <- 20
nu <- 0.5
lambda <- 2.
a_rfx <- 1.
b_rfx <- 1.
a_leaf <- 1.
b_leaf <- 1.
result <- xbart(y, X, omega, rfx_basis, rfx_labels, num_samples, num_burnin, num_trees, nu, lambda, a_leaf, b_leaf, a_rfx, b_rfx, cutpoint_grid_size, random_seed = random_seed)
}
